+++ geostore/tests/test_models/test_layers.py
@@ -8,9 +8,30 @@ from rest_framework.test import APIClient
 
 from geostore.models import Layer
 from geostore.tests.factories import (FeatureFactory, LayerFactory,
-                                      UserFactory)
+                                      LayerSchemaFactory, UserFactory)
 from geostore.tests.utils import get_files_tests
 from geostore.transformations import set_geometry_from_options
+
+
+class LayerModelTestCase(TestCase):
+    def setUp(self):
+        self.layer_schema = LayerSchemaFactory()
+
+    def test_get_property_title_defined(self):
+        """ method should return property title """
+        self.assertEqual(self.layer_schema.get_property_title('age'),
+                         self.layer_schema.schema['properties']['age']['title'])
+
+    def test_get_property_title_undefined(self):
+        """ method should return property name """
+        prop = 'name'
+        self.assertEqual(self.layer_schema.get_property_title(prop),
+                         prop)
+
+    def test_get_property_type(self):
+        """ method should return property type """
+        self.assertEqual(self.layer_schema.get_property_type('age'),
+                         self.layer_schema.schema['properties']['age']['type'])
 
 
 class LayerFromCSVDictReaderTestCase(TestCase):